requirements.txt
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djoser==2.1.0
psycopg2_binary==2.9.3

django-extensions==3.1.5
django-organizations==2.0.1